outline written
The project mainly consists of 4 modules, bot, Game, scoreboard, & lookups. main is only used to instantiate and run the bot once. preparing_dataset if the notebook responsible for preparing the dataset, transforming it from a list of words, into a dictionary of combination of letters as key, and list of words as value